Unrefactor networks
