--- net.go
@@ -1,48 +1,0 @@
-// Copyright 2019, 2021, 2022 The Alpaca Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"log"
-	"net"
-)
-
-func networks(hostname string) []string {
-	if hostname == "" {
-		return []string{"tcp"}
-	}
-	addrs, err := net.LookupIP(hostname)
-	if err != nil {
-		log.Fatal(err)
-	}
-	nets := make([]string, 0, 2)
-	ipv4 := false
-	ipv6 := false
-	for _, addr := range addrs {
-		// addr == net.IPv4len doesn't work because all addrs use IPv6 format.
-		if addr.To4() != nil {
-			ipv4 = true
-		} else {
-			ipv6 = true
-		}
-	}
-	if ipv4 {
-		nets = append(nets, "tcp4")
-	}
-	if ipv6 {
-		nets = append(nets, "tcp6")
-	}
-	return nets
-}
